[issue-566] refactor Pravega options
